fuelgauge: add back battery level animation
This adds the battery level animation when entering battery
settings menu. This was dropped in AOSP, but still exists on Pixel's
(At least in DP4).

When a user enters the battery settings menu, the battery icon will
animate
into the current battery level, as well as the battery level text.

This is a revert of: 4a7c8d7
This is a cherry-pick of: a3c528f

Updated for ODR3 (8.1) by @Sirhc95

Change-Id: I8f64d7eb96d8fb64c614d23272e948d7a70bd5c8
Signed-off-by: HeroBuxx <herobuxx@gmail.com>